# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MontoCRM API
  version: 2.0.0
  description: |
    AI 原生 B2B 大客户 CRM — Phase 2（P5–P9）契约。

    实现对照：
    - P5 Guard hybrid + `/admin/llm/guard/test`；拦截审计 `guard.block`
    - P6 Langfuse Mock + `/metrics` + Ragas eval
    - P7 混合检索 + `/admin/rag/test-retrieval` / reindex
    - P8 钉钉自定义机器人 + `/admin/notify/dingtalk`
    - P9 CI、Nginx TLS/限流、安全头；`/auth/token` 与 `/activities/extract` 可返回 429

    运行时 FastAPI `version` 为 2.0.0。登录实现为 OAuth2 password（`/auth/token`，`/auth/login` 为别名）。
  contact:
    name: MontoCRM
servers:
  - url: http://localhost:8000
    description: Local API（无 /api 前缀；经 Nginx 时为 https://host/api/）
  - url: https://localhost/api
    description: Compose Nginx TLS
...
paths:
  /auth/token:
    post:
      tags: [Auth]
      summary: 登录（表单 username=邮箱）
      operationId: loginToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, description: 邮箱 }
                password: { type: string, format: password }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 凭据错误
        '429':
          description: 过频（默认生产 5 req/min/IP；Nginx 另有 limit_req）
...
  /opportunities:
    get:
      tags: [Opportunities]
      summary: 商机列表
      operationId: listOpportunities
      security:
        - bearerAuth: []
      parameters:
        - name: health_status
          in: query
          schema:
            $ref: '#/components/schemas/HealthStatus'
        - name: stage
          in: query
          schema:
            $ref: '#/components/schemas/OpportunityStage'
        - $ref: '#/components/parameters/PageParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityListResponse'
    post:
      tags: [Opportunities]
      summary: 创建商机
      operationId: createOpportunity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityCreate'
      responses:
        '201':
          description: Created
  /opportunities/{opportunity_id}:
    get:
      tags: [Opportunities]
      summary: 商机详情
      operationId: getOpportunity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
    patch:
      tags: [Opportunities]
      summary: 更新商机（含阶段推进）
      operationId: updateOpportunity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpportunityUpdate'
      responses:
        '200':
          description: OK
...
  /activities/extract:
    post:
      tags: [Activities, Writeback]
      summary: 提交会议纪要并触发 Orchestrator 并行分析（闭环 A）
      operationId: extractFromNotes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractRequest'
      responses:
        '200':
          description: 已生成写回建议，等待确认
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WritebackProposal'
        '429':
          description: 抽取过频（生产 10 req/min/用户；Nginx 另有 limit_req）
...
  /admin/rag/test-retrieval:
    post:
      tags: [Admin RAG]
      summary: 测试 RAG 检索（Embedding + Rerank，不调用 Chat LLM）
      operationId: testRagRetrieval
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagTestRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RagTestResponse'
  /admin/rag/reindex:
    post:
      tags: [Admin RAG]
      summary: 触发全库 memory_chunks re-index（Embedding 变更后）
      operationId: triggerReindex
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Re-index 任务已排队
  /admin/rag/reindex-status:
    get:
      tags: [Admin RAG]
      summary: 查询 re-index 进度
      operationId: getReindexStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReindexStatus'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OpportunityId:
      name: opportunity_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
...
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
...
    HealthStatus:
      type: string
      enum: [green, yellow, red]
    OpportunityStage:
      type: string
      enum: [qualified, discovery, proposal, negotiation, closed_won, closed_lost]
...
    Opportunity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          type: number
        stage:
          $ref: '#/components/schemas/OpportunityStage'
        health_score:
          type: integer
        health_status:
          $ref: '#/components/schemas/HealthStatus'
        competitor:
          type: string
        pain_points:
          type: array
          items:
            type: string
    OpportunityCreate:
      type: object
      required: [account_id, name]
      properties:
        account_id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          type: number
        stage:
          $ref: '#/components/schemas/OpportunityStage'
    OpportunityUpdate:
      type: object
      properties:
        stage:
          $ref: '#/components/schemas/OpportunityStage'
        amount:
          type: number
        competitor:
          type: string
        pain_points:
          type: array
          items:
            type: string
    OpportunityListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Opportunity'
        total:
          type: integer
...
    ExtractRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 50
        account_id:
          type: string
          format: uuid
        opportunity_id:
          type: string
          format: uuid
    WritebackProposal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        opportunity_updates:
          type: object
        contact_updates:
          type: array
          items:
            type: object
        new_contacts:
          type: array
          items:
            type: object
        tasks:
          type: array
          items:
            type: object
        reasoning:
          type: string
        evidence:
          type: array
          items:
            type: object
...
    RagTestRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        account_id:
          type: string
          format: uuid
        opportunity_id:
          type: string
          format: uuid
    RagTestResponse:
      type: object
      properties:
        chunks:
          type: array
          items:
            type: object
            properties:
              chunk_id:
                type: string
                format: uuid
              snippet:
                type: string
              score:
                type: number
        retrieval_meta:
          type: object
          properties:
            raw_query:
              type: string
            rewritten_query:
              type: string
            chunk_strategy:
              type: string
            embedding_provider:
              type: string
            embedding_model:
              type: string
            rerank_enabled:
              type: boolean
            rerank_provider:
              type: string
            rerank_model:
              type: string
    ReindexStatus:
      type: object
      properties:
        status:
          type: string
          enum: [idle, pending, running, completed, failed]
        progress_pct:
          type: integer
        message:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...